requirements.txt
asarPy==1.0.1
psycopg2-binary
python-dotenv

setuptools==80.9.0
wheel==0.45.1
python-dotenv==1.0.1